+++ src/components/dashboard/deposit/TopUpChangelly.js
@@ -18,7 +18,7 @@ at.apply(this, arguments);
 const getChangellyIframe = identityAddress =>
 	`
 	<iframe 
-	src="https://widget.changelly.com?from=btc,eth,xrp&to=dai&amount=0.007&address=${identityAddress}&fromDefault=btc&toDefault=dai&theme=aqua&merchant_id=unl7fil0on3x48m1&payment_id=&v=2" 
+	src="https://widget.changelly.com?from=btc,eth,xrp&to=dai&amount=0.007&address=${identityAddress}&fromDefault=btc&toDefault=dai&theme=aqua&merchant_id=unl7fil0on3x48m1&payment_id=&v=3" 
 	width="100%" height="600" class="changelly" scrolling="no" 
 	onLoad="${onloadFunction}" 
 	style="min-width: 100%; width: 100px; overflow-y: hidden; border: none">Can't load widget</iframe>