+++ src/selectors/selectors.js
@@ -6,8 +6,3 @@ export const createDeepEqualSelector = createSelectorCreator(
 	defaultMemoize,
 	isEqual
 )
-
-export const creatArrayOnlyLengthChangeSelector = createSelectorCreator(
-	defaultMemoize,
-	(a = [], b = []) => a.length === b.length
-)